Updated on 2025-09-01

+++ domain/tokens/models/src/main/java/com/tangem/domain/tokens/model/Quote.kt
@@ -15,5 +15,5 @@ data class Quote(
     val rawCurrencyId: String,
     val fiatRate: BigDecimal,
     val priceChange: BigDecimal,
-    val values: List<Double>? = null,
+    val values: List<Float>? = null,
 )
